Extract console value models
